(fix) Get tests working
